pyproject.toml
[project]
name = "new-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "openai>=1.109.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
